Mixins, rename 101cookbooks files to not use digits
Can't use import statements with modules that start with digits. Also updated files to use mixin to share item parser
